# docs/swagger.yaml
basePath: /api
definitions:
...
  entities.JSONResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
  entities.Room:
    properties:
      cost:
        type: number
      created_at:
        type: string
      id:
        type: string
      status:
        type: string
      updated_at:
        type: string
      vender_id:
        type: string
    type: object
  entities.RoomPayload:
    properties:
      cost:
        type: string
      status:
        type: string
      vendor:
        type: integer
    type: object
...
info:
  contact:
    email: bicosteve4@gmail.com
    name: Bico Oloo
    url: https://github.com/bicosteve
  description: Booking API to make booking reservations
  title: Booking API
  version: "1.0"
paths:
...
  /api/admin/rooms:
    post:
      consumes:
      - application/json
      description: Receives room payload, validate it then send it to service
      operationId: create-room
      parameters:
      - description: Create room
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/entities.RoomPayload'
      produces:
      - application/json
      responses:
        "201":
          description: '{"msg":"created"}'
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: Admin user create a room
      tags:
      - rooms
  /api/admin/rooms/{room_id}:
    delete:
      consumes:
      - application/json
      description: Receives room_id and deletes the room
      operationId: delete-room
      parameters:
      - description: Room ID to delete
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Room deleted successfully
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: delete a room
      tags:
      - rooms
    put:
      consumes:
      - application/json
      description: Receives room payload, validates it, then updates the room by identified
        room_id
      operationId: update-room
      parameters:
      - description: Room ID to update
        in: path
        name: room_id
        required: true
        type: string
      - description: Room update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/entities.RoomPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Room updated successfully
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: update a room
      tags:
      - rooms
...
  /api/user/rooms:
    get:
      consumes:
      - application/json
      description: Retrieve all rooms and filter using query param
      operationId: get-rooms
      parameters:
      - description: Room ID to filter
        in: query
        name: room_id
        type: string
      - description: Room status to filter
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Single room (if exact match)
          schema:
            $ref: '#/definitions/entities.Room'
        "400":
          description: Bad request, validation error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "404":
          description: Bad request, room not found
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: Get a room by rooms and filter by ID
      tags:
      - rooms
...
schemes:
- http
swagger: "2.0"